fix: use static media device id salt (#75)
There's a bug in Around causing device IDs to change when the user starts the
app sometimes. Chromium uses a salt to randomize IDs and prevent browser
fingerprinting. Electron stores this salt in user preferences which seems to
somehow get deleted.

When the device ID changes, Around is unable to find the user's preferred
devices. Using a static media device id salt avoids this issue entirely.